rtw88: decompose while(1) loop of power sequence polling command
The power polling command is one kind of power sequence commands. It's used
to check hardware situation, and subsequent comamnds will be executed if
hardware is ready. A special case is PCIE must toggle BIT_PFM_WOWL and try
again if first try is failed.

According to above description, we can replace the 'while (1)...loop' by
two statements to do first try and retry. In order to reduce indentation
and easier to understand, move polling part to an individual function,
and then flow will match above description of first paragraph.

Change-Id: I4c712fcea83e973821edc6dc133887784827f457
Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>