rtw88: 8723d: implement ops set_tx_power_index to set tx power
The txagc table is used to map rate_id and txagc register address and
mask, and ops set_tx_power_index uses this table to write TX power to
corresponding register. Since 8723D is a 1x1 2.4G 11n chip, only CCK, OFDM
and HT_MCS 0-7 are listed in the table.

Change-Id: I6aedf36ab9cc628836b55637a96e6a940704a654
Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>