rtw88: select deep PS mode when module is inserted
Add a module parameter to select deep PS mode. And the mode
cannot be changed after the module has been inserted, hence
the file mode for the module parameter is read only (0444).

When the device is probed, driver will check the deep PS mode
with different IC's PS mode suppotability. If none of the
PS mode is matched, the deep PS mode is changed to NONE,
means deep PS is disabled.

Change-Id: I1f5a0393c96ec2d244cec72b90bc3df509a02d64
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>