rtw88: sar: Load static SAR table from ACPI WRDS method
ACPI WRDS method returns static SAR table that contains two chains (RF paths)
and five power limit data for each chain. The limit data are corresponding
to certain ranges of frequency, such as 2.4G band, 5.15~5.35G etc.

The data is in Q.3 notation that is the same with SAR entry function, so
we don't need to convert its quantity.

Change-Id: I889f315688a008d82e0912c03e7abccc62ef69b2
Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>