rtw88: sar: Don't induce to SAR WW if regd WW is specified
Normally, WW SAR is the minimum one of all regulatory domains, but directly
use SAR value if WW is specified as 'regd'. So, if SAR value related to WW is
existing, the order to set SAR is important -- SAR WW must be set in the
last one that will overwrite WW SAR value induced by other domains.
For example, if we have 18/17/16 SAR values corresponding to FCC/ETSI/WW,
the steps and values look like
             FCC    ETSI    WW
    Step 1   18             18
    Step 2   --      17     17
    Step 3   --      --     16
    Final    18      17     16

    * '--' means the field is unchanged

This patch will be merged into "rtw88: sar: add SAR of TX power limit".

Change-Id: Ibd6c88ae12ff9b488213bc45c7ecf3ee0aead87d
Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>