rtw88: avoid holding mutex for cancel_delayed_work_sync()
Driver could possibly be dead-locked while canceling works with
*_sync() with mutex lock held. Those cancel_delayed_work_sync()
functions will wait until the work is done, but if we hold the
lock, they will never acquire the lock.

To prevent this, simply release the lock and acquire again after
the works have been canceled. And to avoid the works being queued
again, check if the device is at RTW_FLAG_RUNNING state, otherwise
just return and do nothing.

Change-Id: Idadedcf20e549e98dd0cccd8a15dcf856e90b79a
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>