rtw88: disable TX-AMSDU on 2.4G band
Some tests shows that using AMSDU to aggregate TCP ACKs to specific
APs will degrade the throughput on 2.4G band in 20MHz bandwidth
(< 10 Mbps, should be ~100 Mbps for 2x2). Also found that there's
barely no negative impact if we disable TX AMSDU on 2.4G to connect
to other APs. So it seems like we can just tell mac80211 to not to
aggregate MSDUs by setting the max aggregation size to 1 when
transmitting on 2.4G band.

Note that we still can TX AMSDU on 5G band and benefit from it by
having 50 ~ 70 Mbps throughput improvement.

Change-Id: If7e8cc4157058d3e79702bfb63bf6a562e1c3790
Signed-off-by: Yan-Hsuan Chuang <yhchuang@realtek.com>